+++ docs/changelog.md
@@ -1,7 +1,7 @@
 
 # v7.1.1.1 Beta (????-??-??)
 
-- Korean translation added, Japanese translation fixed. Thanks to the translation team!
+- Korean and Russian translation added, Japanese translation fixed. Thanks to the translation team!
 - Auto build update.
 
 # v7.1.1.0 (2024-02-03)